Session management service/admin app/requirements.txt
fastapi
uvicorn[standard]
mysql.connector
bcrypt